Adding Center View command for the editor.
Center View is similar to emacs "recenter-top-bottom", which
scrolls the viewport of the editor according to the cursor
position, eg, either centering the editor on the cursor;
scrolling the viewport so the cursor is in the top; or
scrolling it so the viewport is in the bottom.

Successive calls to ctrl+g cycles through these actions.