Fix: packages with equal PackageIdentifiers.
BUG:
* Implementation up until now allowed such cases, but did not work
very well with the Browser panel - e.g., when two packages with same
name and version number had a module 'Amod', only the module from
one of the packages would get cached; all attempts to access it
from the Browser pane1 would result in opening the cached module,
even if it was inactive or perhaps removed from the IDE. So the
user might end up editing the wrong 'Amod' and see wrong results
when trying to building.

FIX:
* Observable behavior: preference is given to finding modules from
the active package.
* Implementation: rebuilds the workspace metadata after every time
time it activates a package.
* Sure to work only with Activate Package for now; still have not
tested with "(Add|Remove) Package".

TODO:
* Test some more.
* Probably a change some bits for it to work with "(Add|Remove)
Package".
* Clean up the code.

OTHER: have added a ' ' to a line in the pt_BR.po. It was missing.